+++ packages/api/src/routes/announcements.ts
@@ -4,7 +4,7 @@
  */
 
 import { Hono } from 'hono'
-import { zValidator } from '@hono/zod-validator'
+import { zValidator } from '../middleware/validator'
 import { z } from 'zod'
 import { AnnouncementService } from '../services/announcement-service'
 import {